Add "Wizard service" to modify Wizard state during interactive build
Say I build a library and then discover that it actually depends
on Zlib. At the moment, I have to go back to the start of the
wizard and declare my additional dependency. It would be nice
to just request that the wizard add it from inside the environment.
This is what this PR implements by exposing an AF_UNIX server that
responds to requests from the sandbox.
With a suitable `bb` command, the following works from inside the sandbox:
```
sandbox:${WORKSPACE}/srcdir # bb add Zlib_jll
   Updating registry at `~/.julia/registries/General`
   Updating git-repo `https://github.com/JuliaRegistries/General.git`
  Resolving package versions...
   Updating `/tmp/jl_q0IePp/mRAcrdQ3/.project/Project.toml`
  [83775a58] + Zlib_jll v1.2.11+9
   Updating `/tmp/jl_q0IePp/mRAcrdQ3/.project/Manifest.toml`
  [83775a58] + Zlib_jll v1.2.11+9
  [2a0f44e3] + Base64
  [ade2ca70] + Dates
  [b77e0a4c] + InteractiveUtils
  [76f85450] + LibGit2
  [8f399da3] + Libdl
  [56ddb016] + Logging
  [d6f4376e] + Markdown
  [44cfe95a] + Pkg
  [de0858da] + Printf
  [3fa0cd96] + REPL
  [9a3f8284] + Random
  [ea8e919c] + SHA
  [9e88b42a] + Serialization
  [6462fe0b] + Sockets
  [cf7118a7] + UUIDs
  [4ec0a83e] + Unicode
sandbox:${WORKSPACE}/srcdir #
```

Ideally, that `bb` command, could just be a simple wrapper over
netcat. However, that runs into https://github.com/JuliaLang/julia/issues/35442.
I will add a suitable pure-C version of that BB command to the RootFS
instead. This is just the BinaryBuilder version.